server/requirements.txt
aiohttp>=3.9.0
pywinpty>=2.0.0; sys_platform == 'win32'
